ready test

+++ src/main/java/Desk.kt
@@ -1,9 +1,9 @@
-class Desk(val x:Int, var y:Int, val length:Int, val weight: Int){
-    fun step (limitY:Int,radious:Int, ddy:Int) {
-        y+=ddy
-        val limitMin= 2*(radious+1)
+class Desk(val x: Double, var y: Double, val length: Int, val weight: Int) {
+    fun step(limitY: Double, radious: Double, ddy: Int) {
+        y += ddy
+        val limitMin = 2 * (radious + 1)
         val limitMax = limitY - length - limitMin
-        if (y<=limitMin) y = limitMin
-        if (y>=limitMax) y = limitMax
+        if (y <= limitMin) y = limitMin
+        if (y >= limitMax) y = limitMax
     }
 }
